Fix file size check in JFlexMojoTest.java
The test failed because the file was too big. Fix this by first shrinking
a bit the generated code, and increasing slightly the bound tested: the
generated code has to be more verbose in some places, because many
operations like f(v++) had to be rewritten f(v); v+=1; for scala
portability, and the new generated code has many more "final" keywords.